requirements.txt
groq
pyttsx3
SpeechRecognition
PyAudio
requests
beautifulsoup4
pyautogui
pywhatkit
wikipedia
pynput
python-dotenv
